% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_all_functions.R
\docType{data}
\name{acoustic_subset}
\alias{acoustic_subset}
\title{Phonetics acoustic data (subset)}
\format{
A data frame with 656 rows and 11 variables
}
\description{
A small subset of the phonetics acoustic
data set \code{\link[sparseFLMM]{acoustic}} with observations from two speakers and two items only.
This will not produce meaningful results but can be used as a toy
data set when testing the code.
The variables are as in the full data set, see \code{\link[sparseFLMM]{acoustic}}.
}
\references{
Pouplier, Marianne and Hoole, Philip (2016): Articulatory and
Acoustic Characteristics of German Fricative Clusters,
Phonetica, 73(1), 52--78.

Cederbaum, Pouplier, Hoole, Greven (2016): Functional Linear Mixed Models
for Irregularly or Sparsely Sampled Data. Statistical Modelling, 16(1), 67-88.
}
\keyword{datasets}
