% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_constraint_constructor.R
\name{Predict.matrix.symm.smooth}
\alias{Predict.matrix.symm.smooth}
\title{Predict matrix method for (skew-)symmetric bivariate smooths.}
\usage{
\method{Predict.matrix}{symm.smooth}(object, data)
}
\arguments{
\item{object}{is a \code{symm.smooth} object created by \code{\link{smooth.construct.symm.smooth.spec}},
see \code{\link[mgcv]{smooth.construct}}.}

\item{data}{see \code{\link[mgcv]{smooth.construct}}.}
}
\description{
Predict matrix method for (skew-)symmetric bivariate smooths.
}
\seealso{
\code{\link[mgcv]{Predict.matrix}} and \code{\link[mgcv]{smoothCon}} for details on constructors.
}
\author{
Jona Cederbaum, Almond Stoecker
}
